Improve course user listing and relationships

+++ resources/views/course_users/table.blade.php
@@ -2,44 +2,55 @@
     <table class="table" id="courseUsers-table">
         <thead>
         <tr>
-            <th>User Id</th>
-        <th>Course Id</th>
-        <th>User Account Id</th>
+        <th>Course</th>
         <th>Paid Date</th>
         <th>Expiry Date</th>
         <th>Plan</th>
         <th>Paid Amount</th>
         <th>Status</th>
-            <th colspan="3">Action</th>
+            <th colspan="3"></th>
         </tr>
         </thead>
         <tbody>
         @foreach($courseUsers as $courseUser)
             <tr>
-                <td>{{ $courseUser->user_id }}</td>
-            <td>{{ $courseUser->course_id }}</td>
-            <td>{{ $courseUser->user_account_id }}</td>
-            <td>{{ $courseUser->paid_date }}</td>
-            <td>{{ $courseUser->expiry_date }}</td>
+            <td><a class="text-bold" href="{{route('courses.contents', ['course_id' =>$courseUser->course_id, 'contents'=> 'yes']) }}">
+                </a>
+            </td>
+            <td>
+                @if ($courseUser->paid_date != null)
+                    {{ $courseUser->paid_date -> format('D d M Y') }}
+                @endif
+            </td>
+            <td> 
+                @if ($courseUser->expiry_date != null)
+                    {{ $courseUser->expiry_date -> format('D d M Y') }}
+                @endif
+            </td>
             <td>{{ $courseUser->plan }}</td>
-            <td>{{ $courseUser->paid_amount }}</td>
-            <td>{{ $courseUser->status }}</td>
-                <td width="120">
-                    {!! Form::open(['route' => ['courseUsers.destroy', $courseUser->id], 'method' => 'delete']) !!}
+            <td>#{{ number_format ($courseUser->paid_amount) }}</td>
+            <td>
+            @if ($courseUser->status == 1)
+                active
+            @else
+                inactive
+            @endif
+            </td>
+    <td>
+        @if(Auth::check() AND Auth::user()->role_id < 3)
+    
+            {!! Form::open(['route' => ['courseUsers.destroy', $courseUser->id], 'method' => 'delete']) !!}
                     <div class='btn-group'>
-                        <a href="{{ route('courseUsers.show', [$courseUser->id]) }}"
-                           class='btn btn-default btn-xs'>
-                            <i class="far fa-eye"></i>
-                        </a>
-                        <a href="{{ route('courseUsers.edit', [$courseUser->id]) }}"
-                           class='btn btn-default btn-xs'>
+                       
+                        <a href="{{ route('courseUsers.edit', [$courseUser->id]) }}" class='btn btn-default btn-xs'>
                             <i class="far fa-edit"></i>
                         </a>
-                        {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                        {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
                     </div>
                     {!! Form::close() !!}
-                </td>
-            </tr>
+        @endif
+    </td>
+  </tr>
         @endforeach
         </tbody>
     </table>
